+++ modules/bridge/src/main/ts/ephox/bridge/core/DataProcessors.ts
@@ -8,6 +8,7 @@ import { colorPickerDataProcessor } from '../components/dialog/ColorPicker';
 import { customEditorDataProcessor } from '../components/dialog/CustomEditor';
 import { dropZoneDataProcessor } from '../components/dialog/Dropzone';
 import { iframeDataProcessor } from '../components/dialog/Iframe';
+import { imagePreviewDataProcessor } from '../components/dialog/ImagePreview';
 import { inputDataProcessor } from '../components/dialog/Input';
 import { listBoxDataProcessor } from '../components/dialog/ListBox';
 import { selectBoxDataProcessor } from '../components/dialog/SelectBox';
@@ -32,6 +33,7 @@ const dataProcessors = {
   dropzone: dropZoneDataProcessor,
   input: inputDataProcessor,
   iframe: iframeDataProcessor,
+  imagepreview: imagePreviewDataProcessor,
   selectbox: selectBoxDataProcessor,
   sizeinput: sizeInputDataProcessor,
   slider: sliderInputDataProcessor,